backend/requirements.txt

Django==2.2.7
django-allauth==0.40.0
django-cors-headers==3.1.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.1.0

dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
pyjwt==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
psycopg2-binary==2.8.4
django-heroku==0.3.1